+++ src-tauri/src/lib_commands/app_api/mod.rs
@@ -4,6 +4,7 @@ mod cli_bridge;
 mod core;
 mod navidrome;
 mod perf;
+mod platform;
 mod remote;
 mod integration;
 mod analysis;
@@ -12,6 +13,7 @@ pub(crate) use cli_bridge::*;
 pub(crate) use core::*;
 pub(crate) use navidrome::*;
 pub(crate) use perf::*;
+pub(crate) use platform::*;
 pub(crate) use remote::*;
 pub(crate) use integration::*;
 pub(crate) use analysis::*;
